requirements.txt
# # tensorflow==2.12.0
# # numpy<2
# # flask
# # gunicorn
# matplotlib

# # Flask==2.3.2
# # gunicorn==20.1.0
# # tensorflow==2.12.0
# # numpy<2
# gdown==4.7.1


# tensorflow==2.12.0
# keras==2.12.0
# numpy<2
# flask
# gunicorn
# pillow


tensorflow>=2.16.0
keras>=3.0.0
numpy<2
flask
gunicorn
pillow
matplotlib
gdown==4.7.1